Added utility functions and chnaged configuration for the debugger
This commit adds a new file called utils with all function that I
felt like were missing. I know there are like vim functions that
determine on which os I'm currently on but this was easier for me
because `vim.fn.has('win32')` is to vague at least for me.

Also I added a function which returns the system seperator which
was heavily inspired by nodejs.

And last I added a function which returns the path to a specified
plugin installed with packer.

This was all necessary for the nvim-dap configuration because I
had some errors in the dap repl about not finding files in a
path that doesn't even exist so I thought this was due to a false
dap adapter path for microsofts vscode-js-dap debugger. Turns out
the errors still remain, but it still works so I keep it the way
it turned out.